Rename package to go-htmx-site and move stuff to utils
